+++ effect_subject/admin/signs_and_symptoms_admin.py
@@ -4,7 +4,6 @@ from django_audit_fields.admin import audit_fieldset_tuple
 from ..admin_site import effect_subject_admin
 from ..forms import SignsAndSymptomsForm
 from ..models import SignsAndSymptoms
-from .fieldsets import reporting_fieldset_tuple
 from .modeladmin import CrfModelAdmin
 
 
@@ -16,27 +15,50 @@ class SignsAndSymptomsAdmin(CrfModelAdmin):
     fieldsets = (
         (None, {"fields": ("subject_visit", "report_datetime")}),
         (
-            "CM signs and symptoms",
-            {"fields": ("cm_signs_symptoms",)},
-        ),
-        (
             "Signs and symptoms",
             {
                 "fields": (
-                    "signs_and_symptoms",
+                    "any_sx",
+                    "current_sx",
+                    "current_sx_other",
+                    "reportable_as_ae",
+                    "current_sx_gte_g3",
+                    "current_sx_gte_g3_other",
                     "headache_duration",
+                    # "headache_duration_microseconds",
                     "visual_field_loss",
                 )
             },
         ),
-        reporting_fieldset_tuple,
+        (
+            "Reporting",
+            {"fields": ("patient_admitted",)},
+        ),
+        (
+            "CM signs and symptoms",
+            {
+                "fields": (
+                    "cm_sx",
+                    "cm_sx_lp_done",
+                    "cm_sx_bloods_taken",
+                    "cm_sx_patient_admitted",
+                )
+            },
+        ),
         audit_fieldset_tuple,
     )
 
-    filter_horizontal = ["signs_and_symptoms"]
+    filter_horizontal = [
+        "current_sx",
+        "current_sx_gte_g3",
+    ]
 
     radio_fields = {
-        "cm_signs_symptoms": admin.VERTICAL,
+        "any_sx": admin.VERTICAL,
+        "cm_sx": admin.VERTICAL,
         "reportable_as_ae": admin.VERTICAL,
         "patient_admitted": admin.VERTICAL,
+        "cm_sx_lp_done": admin.VERTICAL,
+        "cm_sx_bloods_taken": admin.VERTICAL,
+        "cm_sx_patient_admitted": admin.VERTICAL,
     }