+++ effect_subject/baker_recipes.py
@@ -2,11 +2,19 @@ from edc_visit_tracking.constants import SCHEDULED
 from faker import Faker
 from model_bakery.recipe import Recipe
 
-from .models import Followup, StudyTreatment, SubjectRequisition, SubjectVisit
+from .models import (
+    Followup,
+    SignsAndSymptoms,
+    StudyTreatment,
+    SubjectRequisition,
+    SubjectVisit,
+)
 
 fake = Faker()
 
 followup = Recipe(Followup)
+
+signsandsymptoms = Recipe(SignsAndSymptoms)
 
 studytreatment = Recipe(StudyTreatment)